requirements.txt
alembic==1.4.2
Cerberus==1.3.2
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-Testing==0.8.0
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-JSON==0.3.3
